Add a test for the Tour Guide NPC that will be added to the game.
I've added a test to the semos city directory that will instantiate the
NPC and check all the necessary behaviour for a tour (all of release 1
WBS). However, to achieve this I had to implement a simple NPC without
any of the tour functionality so that it could be insantiated.