Increase Bag inventory slots to 16 from 12 and added chair item along with a test for chair item creation
When applied, this commit will increase the existing bag inventory slots
from 12 to 16. This was quite simple as was changing an existing
settings.In addition this commit will implement an item known as a
chair into the game by adding to the misc items xml file and the daily
quest to generate it within the game.Futhermore,this commit comes with
a test which will check the creation of the item known as chair.